movees color swatches to overview page removes unnecessary intro text on palette page sets order of palette colors from 100-600 sets open state to default color (400 levels and white)
